Update README completley
This is a huge overhaul of the versions prior, but it is not what a v3 would have originally looked like. There wasn't much change to this between v1, v2, or v3. However, 2v2 does bring this. Because I had to go BACK in time, I wanted to update a lot of stuff.
Even the license will be updated.

I had help from bolt.new, as I know virtually nothing about github markdown.